Moved channels last custom ops into individual files.
